+++ src/magpylib/_src/obj_classes/class_current_Polyline.py
@@ -5,6 +5,8 @@
 import warnings
 from typing import ClassVar
 
+import numpy as np
+
 from magpylib._src.display.traces_core import make_Polyline
 from magpylib._src.exceptions import MagpylibDeprecationWarning
 from magpylib._src.fields.field_BH_polyline import current_vertices_field
@@ -41,6 +43,12 @@ class Polyline(BaseCurrent):
 
     current: float, default=`None`
         Electrical current in units of A.
+
+    volume: float
+        Read-only. Object physical volume in units of m^3.
+
+    centroid: np.ndarray, shape (3,) or (m,3)
+        Read-only. Object centroid in units of m.
 
     parent: `Collection` object or `None`
         The object is a child of it's parent collection.
@@ -97,7 +105,7 @@ class Polyline(BaseCurrent):
         # init inheritance
         super().__init__(position, orientation, current, style, **kwargs)
 
-    # property getters and setters
+    # Properties
     @property
     def vertices(self):
         """
@@ -118,6 +126,17 @@ class Polyline(BaseCurrent):
         if self.vertices is None:
             return "no vertices"
         return f"{unit_prefix(self.current)}A" if self.current else "no current"
+
+    # Methods
+    def _get_volume(self):
+        """Volume of object in units of m³."""
+        return 0.0
+
+    def _get_centroid(self):
+        """Centroid of object in units of m."""
+        if self.vertices is not None:
+            return np.mean(self.vertices, axis=0) + self.position
+        return self.position
 
 
 class Line(Polyline):